test(streams): Refactor type tests.
